switching to axios with tests
